perf(api): split parallel review tests by subject
node:test schedules database test files in parallel but serializes every case within a file. Extract the shared fixture and separate revalidation, specification-read, and join behavior so the gate can schedule those subjects independently without dropping a case or assertion.
